# Conflicts:
#	media/gpu/mac/vt_config_util.mm
#	testing/buildbot/test_suite_exceptions.pyl

+++ components/segmentation_platform/internal/selection/segment_result_provider.h
@@ -1,4 +1,4 @@
-// Copyright 2021 The Chromium Authors. All rights reserved.
+// Copyright 2021 The Chromium Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -50,14 +50,14 @@ class SegmentResultProvider {
   struct SegmentResult {
     explicit SegmentResult(ResultState state);
     SegmentResult(ResultState state,
-                  int rank,
+                  float rank,
                   std::unique_ptr<ModelExecutionResult> execution_result);
     ~SegmentResult();
     SegmentResult(SegmentResult&) = delete;
     SegmentResult& operator=(SegmentResult&) = delete;
 
     ResultState state = ResultState::kUnknown;
-    absl::optional<int> rank;
+    absl::optional<float> rank;
 
     // The execution result is only available when the model is executed.
     // TODO(ssid): Support storing inputs to disk if needed.
